desafio testeo funcionalidades

+++ containers/contenedor.mongodb.js
@@ -55,6 +55,12 @@ class ContenedorMongodb {
     //Eliminar mensaje por su id.
     async deleteById(idItem) {
         try {
+            const itemAEliminar = await this.coleccion.find({_id: {$eq: idItem}})
+            if(itemAEliminar === []){
+                console.log(itemAEliminar)
+                throw new Error('no existe el producto')
+            }
+            
             await this.coleccion.deleteOne({_id: {$eq: idItem}});
         } catch (error) {
             logError.error(error);
